CS-3220 Mid term 1
CS3220 Midterm
Please deploy your application on CS3, and upload all source files to CSNS. The source files should include all the source code, documentation (optional), and an HTML file midterm.html which contains a hyperlink to your application on the CS3 server. Note that file uploading will be disabled automatically after the due time, and late submission will not be accepted.

In this exam you are going to develop a web application to help Help Desk schedule appointments for customers. Your application  must use the MVC architecture as discussed in class, and the application must meet the following requirements:

Servlets cannot be used to generate HTML content.
No scripting elements are allowed in JSP pages.
1. Display Appointments Page

The application should have a Appointments page that lists all the customers that are currently managed by the application. For example:

Appointment Time	Scheduled For	Department	Status
2022-10-27 11:30 AM	John Doe	Computers	Completed	Update
2022-10-27 1:30 PM	Jane Doe	Video Games	Scheduled	Update
2022-10-27 3:00 PM	Jack Smith	Appliances	Assinged	Update
2022-10-27 3:00 PM	Jack Smith	Movies	Assinged	Update
Create Appointment

We assume that there are four Departments: Computers, Video Games, Appliances and Movies. The status of the appointment can be Scheduled, Cancelled, Assinged or Completed. When an appointment is created, its status is Scheduled. An ITS Help Desk personnel can assign the appointment to a technician and the status of the appointment should change to Assigned. After the technician completes the appointment, he or she can update the appointment to set its status to Completed.

Note that Appointment Time should be formatted as shown in the example above.

2. Create Appointments

Your Name:
Appointment Time:
11:00 AM
Department:	ComputersVideo GamesAppliancesMovies
Reason:
Clicking on the Create Service Request link should take a user to the Create Request page where a new service request can be created. After the user fills out the form and clicks the Create button, the application should create a service request and redirect the user back to the Requests page.

3. Update Appointment Status

A user, presumably an ITS supervisor or technician, can click on the Update link of an appointment to update the status of the appointment. If the status of an appointment is Scheduled, an ITS supervisor can assign the appointment to one of the three ITS technician: Amy, Bob, and Tom. For example:

Computers Service requested by Jane Doe at 2022-10-27 1:30 PM.

Status: Scheduled

Reason:

Looking for a new laptop.

Assign this request to
Amy

If the status of a request is Assigned, the ITS technician who completed the appointment can change the status to Completed. For example:

Computers Service requested by Jane Doe at 2022-10-27 1:30 PM.

Status: Assigned to Bob

Reason:

Looking for a new laptop.

Change the status to Completed? Yes No Cancelled

Clicking on Yes should change the status to Completed and redirect the user to the Appointments page; clicking on No simply takes the user to the Appointments page; Clicking on Cancelled should change the status to Cancelled.

If the status of an appointments is Completed or Cancelled, the Update link for that appointments on the Appointments page should be removed.